requirements.txt
streamlit
selenium
webdriver-manager
pandas
numpy
openpyxl 
xlsxwriter
pywin32
pyautogui
pillow
streamlit-lottie
